# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Yamo Laundry API
  version: 1.0.1
  description: >-
    REST API for the Yamo laundry platform. Includes customer, orders,
    tasks, billing, geo, notifications, livreur management and live tracking.
servers:
  - url: http://localhost:8080
    description: Local
  - url: https://api.yamo.example
    description: Production
...
paths:
...
  /api/orders:
    get:
      tags: [Orders]
      summary: List orders by customer
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: customerId
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Order' }
    post:
      tags: [Orders]
      summary: Create order
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderRequest'
      responses:
        '201':
          description: Created
          headers:
            Location:
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
  /api/orders/{id}:
    get:
      tags: [Orders]
      summary: Get order
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
        '404': { description: Not found }
  /api/orders/{id}/status:
    post:
      tags: [Orders]
      summary: Update order status
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StatusUpdateRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
  /api/orders/{id}/pickup:
    post:
      tags: [Orders]
      summary: Schedule pickup
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PickupRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Pickup' }
...
  /api/tasks:
    get:
      tags: [Tasks]
      summary: List tasks in a time window
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: from
          required: true
          schema: { type: string, format: date-time }
        - in: query
          name: to
          required: true
          schema: { type: string, format: date-time }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Task' }
    post:
      tags: [Tasks]
      summary: Create task
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: customerId
          required: false
          schema: { type: integer, format: int64 }
        - in: query
          name: orderId
          required: false
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Task' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Task' }
  /api/tasks/{taskId}/assign-livreur:
    post:
      tags: [Tasks]
      summary: Assign livreur to task
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: taskId
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Livreur' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Task' }
  /api/tasks/{taskId}/status:
    post:
      tags: [Tasks]
      summary: Update task status
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: taskId
          required: true
          schema: { type: integer, format: int64 }
        - in: query
          name: status
          required: true
          schema: { $ref: '#/components/schemas/TaskStatus' }
        - in: query
          name: notes
          required: false
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Task' }
  /api/tasks/livreur/{livreurId}:
    get:
      tags: [Tasks]
      summary: List tasks for a livreur within a time window
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: livreurId
          required: true
          schema: { type: integer, format: int64 }
        - in: query
          name: from
          required: true
          schema: { type: string, format: date-time }
        - in: query
          name: to
          required: true
          schema: { type: string, format: date-time }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Task' }
  /api/tasks/customer/{customerId}:
    get:
      tags: [Tasks]
      summary: List tasks for a customer
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: customerId
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Task' }
  /api/livreurs:
    get:
      tags: [Livreurs]
      summary: List livreurs
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: active
          required: false
          schema: { type: boolean }
        - in: query
          name: vehicleType
          required: false
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Livreur' }
    post:
      tags: [Livreurs]
      summary: Create livreur
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Livreur' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Livreur' }
  /api/livreurs/{id}:
    get:
      tags: [Livreurs]
      summary: Get livreur
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Livreur' } } } }
        '404': { description: Not found }
    put:
      tags: [Livreurs]
      summary: Update livreur
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Livreur' }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Livreur' } } } }
    delete:
      tags: [Livreurs]
      summary: Delete livreur
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '204': { description: No content }
...
  /api/notifications/whatsapp/test:
    post:
      tags: [Notifications]
      summary: Send a WhatsApp text message (test)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                to: { type: string, description: E.164 number }
                text: { type: string }
      responses:
        '200': { description: OK }
  /api/notifications/whatsapp/webhook:
    get:
      tags: [Notifications]
      summary: Verify WhatsApp webhook
      parameters:
        - in: query
          name: hub.mode
          schema: { type: string }
        - in: query
          name: hub.verify_token
          schema: { type: string }
        - in: query
          name: hub.challenge
          schema: { type: string }
      responses:
        '200': { description: Challenge echoed }
        '403': { description: Forbidden }
    post:
      tags: [Notifications]
      summary: Receive WhatsApp webhook events
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        '200': { description: OK }
  /api/notifications/whatsapp/campaign:
    post:
      tags: [Notifications]
      summary: Send WhatsApp campaign to a targeted segment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CampaignReq'
      responses:
        '200': { description: OK }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CustomerRef:
      type: object
      properties:
        id: { type: integer, format: int64 }
...
    Order:
      type: object
      properties:
        id: { type: integer, format: int64 }
        customer: { $ref: '#/components/schemas/CustomerRef' }
        items:
          type: array
          items: { $ref: '#/components/schemas/OrderItem' }
        pickup: { $ref: '#/components/schemas/Pickup' }
        delivery: { $ref: '#/components/schemas/Delivery' }
        status: { type: string }
        totalAmount: { type: number, format: double }
        discountAmount: { type: number, format: double }
        createdAt: { type: string, format: date-time }
    OrderItem:
      type: object
      properties:
        id: { type: integer, format: int64 }
        orderId: { type: integer, format: int64 }
        itemType: { type: string }
        quantity: { type: integer }
        pricePerUnit: { type: number, format: double }
        specialInstructions: { type: string }
    OrderRequest:
      type: object
      required: [customerId]
      properties:
        customerId: { type: integer, format: int64 }
        promoCode: { type: string }
        items:
          type: array
          items: { $ref: '#/components/schemas/OrderItemRequest' }
    OrderItemRequest:
      type: object
      properties:
        itemType: { type: string }
        quantity: { type: integer }
        pricePerUnit: { type: number, format: double }
        specialInstructions: { type: string }
    StatusUpdateRequest:
      type: object
      properties:
        status: { type: string }
    Pickup:
      type: object
      properties:
        id: { type: integer, format: int64 }
        orderId: { type: integer, format: int64 }
        contactName: { type: string }
        contactPhone: { type: string }
        address: { type: string }
        livreur: { $ref: '#/components/schemas/LivreurRef' }
        scheduledDate: { type: string, format: date-time }
        actualDate: { type: string, format: date-time }
        status: { type: string }
    Delivery:
      type: object
      properties:
        id: { type: integer, format: int64 }
        orderId: { type: integer, format: int64 }
        contactName: { type: string }
        contactPhone: { type: string }
        address: { type: string }
        livreur: { $ref: '#/components/schemas/LivreurRef' }
        scheduledDate: { type: string, format: date-time }
        actualDate: { type: string, format: date-time }
        status: { type: string }
    PickupRequest:
      type: object
      properties:
        contactName: { type: string }
        contactPhone: { type: string }
        address: { type: string }
        scheduledDate: { type: string, format: date-time }
...
    Task:
      type: object
      properties:
        id: { type: integer, format: int64 }
        type: { $ref: '#/components/schemas/TaskType' }
        order: { $ref: '#/components/schemas/OrderRef' }
        customer: { $ref: '#/components/schemas/CustomerRef' }
        title: { type: string }
        description: { type: string }
        addressLine: { type: string }
        latitude: { type: number, format: double }
        longitude: { type: number, format: double }
        scheduledAt: { type: string, format: date-time }
        remindBeforeMinutes: { type: integer }
        assigneeLivreur: { $ref: '#/components/schemas/LivreurRef' }
        participantsPhones:
          type: array
          items: { type: string }
        status: { $ref: '#/components/schemas/TaskStatus' }
        notes: { type: string }
        proofPhotoUrl: { type: string }
        proofSignatureUrl: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    TaskType:
      type: string
      enum: [PICKUP, DELIVERY, OTHER]
    TaskStatus:
      type: string
      enum: [SCHEDULED, IN_PROGRESS, DONE, CANCELLED, MISSED]
    OrderRef:
      type: object
      properties:
        id: { type: integer, format: int64 }
    Livreur:
      type: object
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        surname: { type: string }
        nationalId: { type: string }
        phone: { type: string }
        email: { type: string }
        vehicleType: { type: string }
        vehicleBrand: { type: string }
        vehicleColor: { type: string }
        vehicleRegistration: { type: string }
        photoUrl: { type: string }
        comment: { type: string }
        active: { type: boolean }
    LivreurRef:
      type: object
      properties:
        id: { type: integer, format: int64 }
...
    CampaignReq:
      type: object
      properties:
        minAmount: { type: number, format: double }
        minOrders: { type: integer, format: int64 }
        periodDays: { type: integer, format: int32 }
        inactiveDays: { type: integer, format: int32 }
        segment: { type: string }
        text: { type: string }
        templateName: { type: string }
        templateLanguage: { type: string }
        templateParams:
          type: array
          items: { type: string }
security:
  - bearerAuth: []